requirements.txt
streamlit
PyPDF2
python-dotenv

langchain==0.1.17
langchain-community==0.0.36
langchain-text-splitters==0.0.1
langchain-google-genai==0.0.9
langchain-huggingface==0.0.3

sentence-transformers
faiss-cpu
